Refactor: host 등록페이지 리팩토링
